requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.30.6
psycopg2-binary==2.9.12
langchain==0.3.7
langchain-openai==0.2.9
langchain-core==0.3.18
pydantic==2.9.2
pydantic-settings==2.5.2
streamlit==1.38.0
faker==30.3.0
python-dotenv==1.0.1
httpx==0.27.2
requests==2.32.3
pandas==2.2.3
